pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.1",
    "fastembed>=0.8.0",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.1.14",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.18",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.8",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]
